[libft] ft_strtrim size_t to int

+++ libft/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: minjinki <minjinki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/09 18:16:00 by minjinki          #+#    #+#             */
-/*   Updated: 2022/07/20 20:20:47 by minjinki         ###   ########.fr       */
+/*   Updated: 2022/09/22 14:04:09 by minjinki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 char	*ft_strtrim(char const *s1, char const *set)
 {
 	char	*res;
-	size_t	front;
-	size_t	rear;
+	int		front; // size_t로 되어 있으면 밑에 rear == -1이 비교 불가능하지.. 바보야
+	int		rear;
 
 	if (!s1 || !set)
 		return (NULL);
